api_backend/requirements-local.txt
# Local development requirements (Windows compatible)
# Simplified version without problematic dependencies

# Core FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Database dependencies (SQLite only for local dev)
sqlalchemy[asyncio]==2.0.23
aiosqlite>=0.19.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt>=4.0.1

# HTTP client and async HTTP
aiohttp==3.9.5
requests==2.31.0
aiofiles==23.2.1

# Google Cloud Platform services (optional for local)
google-auth==2.40.3
google-auth-oauthlib==1.1.0
google-api-python-client==2.110.0
google-cloud-storage==2.16.0
google-generativeai==0.3.2

# AI Services
openai==1.3.5

# Image processing
Pillow==10.3.0

# Additional utilities
python-dateutil==2.8.2
pytz>=2023.3
typing-extensions>=4.8.0
python-magic-bin==0.4.14 